(enhancement) Integrate D3 Graph into CompareContainer:
* Merged chart.jsx (container) into CompareContainer
* Commented out files
* Removed un-needed files
