--- src/components/app.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Component } from 'react';
-
-export default class App extends Component {
-
-  render() {
-    return (
-      <div>
-        
-      </div>
-    );
-  }
-}
