--- src/components/posts_index.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-import Chart from '../containers/chart';
-
-export default class Test extends Component {
-  render() {
-    return (
-      <div>
-        <h1> Hello, world! </h1>
-        <Chart />
-      </div>
-    );
-  }
-}
