+++ src/containers/AppContainer.jsx
@@ -7,7 +7,6 @@ import { Link } from 'react-router';
 
 class Test extends Component {
   render() {
-    // console.log(this.props);
     const { dispatch, isAuthenticated, errorMessage } = this.props;
 
     return (
@@ -17,8 +16,6 @@ class Test extends Component {
           errorMessage={errorMessage}
           dispatch={dispatch}
         />
-        <Link to="/test">Link to test</Link>
-         <br></br>
         <Link to="/offers">Link to offer</Link>
         {this.props.children}
       </div>
@@ -32,8 +29,8 @@ function mapStateToProps(state) {
 
   return {
     isAuthenticated,
-    errorMessage
-  }
+    errorMessage,
+  };
 };
 
-export default connect(mapStateToProps)(Test)
+export default connect(mapStateToProps)(Test);
